% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows.R
\name{format_all_episodes}
\alias{format_all_episodes}
\title{format_all_episodes}
\usage{
format_all_episodes(all_episodes)
}
\arguments{
\item{all_episodes}{}
}
\value{
a data frame that has been cleaned cleaned and filtered of missing values
}
\description{
format_all_episodes
}
\examples{
format_all_episodes(all_episodes)
}
