% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows.R
\name{get_all_episodes}
\alias{get_all_episodes}
\title{get_all_episodes}
\usage{
get_all_episodes(show_id)
}
\arguments{
\item{show_id}{}
}
\value{
a data frame of the parsed json of the get response
}
\description{
get_all_episodes
}
\examples{
get_all_episodes(show_id)
}
